[프로젝트] OCR과 LLM을 활용한 개인 맞춤 영양정보 제공 End-to-End 서비스/Nutrition_Advisor/project/streamlit_UI/requirements.txt
bardapi
streamlit
opencv-python
numpy
easyocr
Pillow
gTTS
paddlepaddle==2.5.0
paddleocr